@@ -5,7 +5,7 @@
     "type": "git",
     "url": "https://github.com/tanstack/form.git"
   },
-  "packageManager": "pnpm@9.12.1",
+  "packageManager": "pnpm@9.12.2",
   "type": "module",
   "scripts": {
     "clean": "pnpm --filter \"./packages/**\" run clean",
@@ -41,34 +41,34 @@
     "@eslint-react/eslint-plugin": "^1.15.0",
     "@solidjs/testing-library": "^0.8.10",
     "@tanstack/config": "^0.13.4",
-    "@testing-library/jest-dom": "^6.5.0",
+    "@testing-library/jest-dom": "^6.6.2",
     "@testing-library/react": "^16.0.1",
     "@testing-library/user-event": "^14.5.2",
     "@testing-library/vue": "^8.1.0",
     "@types/node": "^20.14.10",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
-    "@vitest/coverage-istanbul": "^2.1.2",
-    "eslint": "^9.12.0",
+    "@vitest/coverage-istanbul": "^2.1.3",
+    "eslint": "^9.13.0",
     "eslint-plugin-react-hooks": "^5.1.0-rc-fb9a90fa48-20240614",
     "jsdom": "^25.0.1",
     "knip": "^5.33.3",
-    "nx": "^19.8.4",
+    "nx": "^19.8.6",
     "prettier": "^3.3.3",
     "publint": "^0.2.11",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "rimraf": "^5.0.10",
-    "sherif": "^1.0.0",
+    "sherif": "^1.0.1",
     "solid-js": "^1.9.2",
-    "typescript": "5.4.2",
+    "typescript": "5.4.5",
     "typescript49": "npm:typescript@4.9",
     "typescript50": "npm:typescript@5.0",
     "typescript51": "npm:typescript@5.1",
     "typescript52": "npm:typescript@5.2",
     "typescript53": "npm:typescript@5.3",
-    "vite": "^5.4.8",
-    "vitest": "^2.1.2",
+    "vite": "^5.4.9",
+    "vitest": "^2.1.3",
     "vue": "^3.3.4"
   }
 }
